+++ lib/source_impl.h
@@ -62,10 +62,10 @@ private:
   size_t d_mtu;
   pmt::pmt_t d_message_port;
 
-  float d_frequency;
+  double d_frequency;
   float d_gain;
-  float d_sampling_rate;
-  float d_bandwidth;
+  double d_sampling_rate;
+  double d_bandwidth;
   std::string d_antenna;
   size_t d_nchan;
   gr_complexd d_dc_offset;
@@ -101,7 +101,7 @@ private:
 
 public:
   source_impl(size_t nchan, const std::string device, const std::string devname,
-              const std::string args, float sampling_rate, const std::string type);
+              const std::string args, double sampling_rate, const std::string type);
   ~source_impl();
 
   virtual bool stop();
@@ -139,7 +139,7 @@ public:
    * \param channel an available channel on the device
    * \param frequency center frequency in Hz
    */
-  void set_frequency(size_t channel, float frequency);
+  void set_frequency(size_t channel, double frequency);
 
   /*!
    * Set the center frequency for the specified RX chain of the element.
@@ -150,7 +150,7 @@ public:
    * \param name an available element name
    * \param frequency center frequency in Hz
    */
-  void set_frequency(size_t channel, const std::string &name, float frequency);
+  void set_frequency(size_t channel, const std::string &name, double frequency);
 
   virtual void set_overall_gain(size_t channel, float gain, bool manual_mode);
 
@@ -189,14 +189,14 @@ public:
    * \param channel an available channel on the device
    * \param sample_rate the sample rate samples per second
    */
-  void set_sample_rate(size_t channel, float sample_rate);
+  void set_sample_rate(size_t channel, double sample_rate);
 
   /*!
    * Set the baseband filter width of the RX chain
    * \param channel an available channel on the device
    * \param bandwidth the baseband filter width in Hz
    */
-  void set_bandwidth(size_t channel, float bandwidth);
+  void set_bandwidth(size_t channel, double bandwidth);
 
   /*!
    * Set the antenna element for the RX chain.
